requirements.txt
numpy
pandas
scikit-learn
opencv-python
matplotlib
streamlit
scikit-image
pillow
tqdm
